requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.23
asyncmy==0.2.8
python-dotenv==1.0.1
cryptography==42.0.2
